# Namespace efficiency