@RazrFalcon 2020-04-18
Yes, I also thought about it, but I don't see a good solution for now. It obviously will be a breaking changes and this is not a problem.

My idea is to store namespaces as they appear in the document, and resolve them on request. This will dramatically reduce the memory usage, but I bet it will hit the performance too much. So it's not a solution, really. But I'm open for other ideas.