@RazrFalcon 2020-04-18
I guess a better idea is too find a way to somehow compact the current `Vec<Namespace>`. For example, if we have only one namespace on the root element, the current algorithm will still duplicate it for each node. But technically we can store only the first one and set `0..1` ns range for all elements.

Of course, it will not work when document has a lot of mixed namespaces, but it still a good direction.