@tomjw64 2020-04-18
I am considering simply reducing duplication in this manner, only cloning namespaces when a node's namespaces != it's parent. So in the worst case (a namespace declaration on every node), it would perform as it does currently. Will open another PR soon.