@tomjw64 2020-04-18
Okay #31 I think is a much safer change and does not have the same tradeoffs this branch does.